move cached consumers to separate package
