+++ debezium-server-batch-sinks/src/test/java/io/debezium/server/cachedbatch/cache/InfinispanCacheSimpleTestProfile.java
@@ -6,28 +6,27 @@
  *
  */
 
-package io.debezium.server.batch.cache.infinispan.cachetypes;
+package io.debezium.server.cachedbatch.cache;
 
 import io.quarkus.test.junit.QuarkusTestProfile;
 
 import java.util.HashMap;
 import java.util.Map;
 
-public class InfinispanCacheRocksdbTestProfile implements QuarkusTestProfile {
+public class InfinispanCacheSimpleTestProfile implements QuarkusTestProfile {
 
   //This method allows us to override configuration properties.
   @Override
   public Map<String, String> getConfigOverrides() {
     Map<String, String> config = new HashMap<>();
 
-    config.put("debezium.sink.type", "batch");
-    config.put("debezium.sink.batch.row-limit", "100");
-    config.put("debezium.sink.batch.time-limit", "3000");
-    config.put("debezium.source.max.batch.size", "500");
-    config.put("debezium.source.poll.interval.ms", "10");
+    config.put("debezium.sink.type", "sparkcachedbatch");
+    config.put("debezium.sink.batch.row-limit", "50");
+    config.put("debezium.sink.batch.time-limit", "1000");
 // ==================== SINK = CACHE ====================
+    config.put("debezium.sink.batch.cache", "infinispan");
     config.put("debezium.sink.batch.cache.memory-maxcount", "1254");
-    config.put("debezium.sink.batch.cache.store", "rocksdb");
+    config.put("debezium.sink.batch.cache.store", "simple");
     config.put("debezium.sink.batch.cache.purge-on-startup", "true");
     config.put("debezium.sink.batch.cache.max-batch-size", "1254");
