+++ debezium-server-batch-sinks/src/test/java/io/debezium/server/cachedbatch/cache/MemoryCacheTestProfile.java
@@ -6,25 +6,28 @@
  *
  */
 
-package io.debezium.server.batch.cache;
+package io.debezium.server.cachedbatch.cache;
 
 import io.quarkus.test.junit.QuarkusTestProfile;
 
 import java.util.HashMap;
 import java.util.Map;
 
-public class InfinispanCacheSimpleTestProfile implements QuarkusTestProfile {
+public class MemoryCacheTestProfile implements QuarkusTestProfile {
 
   //This method allows us to override configuration properties.
   @Override
   public Map<String, String> getConfigOverrides() {
     Map<String, String> config = new HashMap<>();
 
-    config.put("debezium.sink.type", "sparkcachedbatch");
-    config.put("debezium.sink.batch.row-limit", "50");
-    config.put("debezium.sink.batch.time-limit", "1000");
+    config.put("debezium.sink.type", "sparkbatch");
+    config.put("debezium.sink.batch.row-limit", "20000");
+    config.put("debezium.sink.batch.time-limit", "3000");
+    config.put("debezium.source.max.batch.size", "20000");
+    config.put("debezium.source.max.queue.size", "40000");
+    config.put("debezium.source.poll.interval.ms", "20000");
 // ==================== SINK = CACHE ====================
-    config.put("debezium.sink.batch.cache", "infinispan");
+    config.put("debezium.sink.batch.cache", "memory");
     config.put("debezium.sink.batch.cache.memory-maxcount", "1254");
     config.put("debezium.sink.batch.cache.store", "simple");
     config.put("debezium.sink.batch.cache.purge-on-startup", "true");
