+++ debezium-server-batch-sinks/src/test/java/io/debezium/server/cachedbatch/cache/infinispan/cachetypes/InfinispanCacheRocksdbTestProfile.java
@@ -6,14 +6,14 @@
  *
  */
 
-package io.debezium.server.batch.cache.infinispan.cachetypes;
+package io.debezium.server.cachedbatch.cache.infinispan.cachetypes;
 
 import io.quarkus.test.junit.QuarkusTestProfile;
 
 import java.util.HashMap;
 import java.util.Map;
 
-public class InfinispanCacheLocalTestProfile implements QuarkusTestProfile {
+public class InfinispanCacheRocksdbTestProfile implements QuarkusTestProfile {
 
   //This method allows us to override configuration properties.
   @Override
@@ -27,7 +27,7 @@ public class InfinispanCacheLocalTestProfile implements QuarkusTestProfile {
     config.put("debezium.source.poll.interval.ms", "10");
 // ==================== SINK = CACHE ====================
     config.put("debezium.sink.batch.cache.memory-maxcount", "1254");
-    config.put("debezium.sink.batch.cache.store", "local");
+    config.put("debezium.sink.batch.cache.store", "rocksdb");
     config.put("debezium.sink.batch.cache.purge-on-startup", "true");
     config.put("debezium.sink.batch.cache.max-batch-size", "1254");
